water-ops-interop-kit/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "water-ops-interop-kit"
version = "0.1.0"
description = "Water Ops Interoperability Kit (WOIK) for water and wastewater systems"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "DaScient"}]
keywords = ["water", "wastewater", "scada", "telemetry", "interoperability", "metrics", "utilities"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Environmental Science",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "pydantic>=2.5",
  "numpy>=1.24",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "rich>=13.6",
  "typer>=0.9",
]

[project.optional-dependencies]
vis = [
  "plotly>=5.18",
  "matplotlib>=3.8",
]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.4.0",
  "black>=24.1",
  "mypy>=1.8",
  "pre-commit>=3.6",
]

[project.scripts]
woik = "woik.cli:app"

[project.entry-points."woik.plugins"]
# Third-party plug-ins can register here.
# Example:
# acme_scada = "acme_woik_plugin.scada:AcmeScadaPlugin"

[tool.setuptools]
packages = {find = {include = ["woik*"]}}

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B", "UP", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
exclude = ["notebooks", "docs", "site", "dist", "build"]
